+++ Makefile
@@ -170,6 +170,7 @@ lint: lint-check-deps
 		--deadline 300s \
 		--exclude 'return value not checked' \
 		--exclude 'possible misuse of unsafe.Pointer' \
+		--exclude 'x \^ 0 always equals x' \
 		./...
 
 lint-check-deps:
